SEAR-30 added LicesingIssue entity
