+++ legacy-data-access-services/src/test/resources/hibernate.cfg.xml
@@ -60,8 +60,8 @@
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.ClientServiceProvider"/>
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.OtherClientName" />
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.TribalMembershipVerification" />
+    <mapping class="gov.ca.cwds.data.legacy.cms.entity.LicensingIssue"/>
     <mapping class="gov.ca.cwds.data.persistence.cms.BaseClientImpl" />
-
 
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ActiveServiceComponentType"/>
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ApprovalStatusType"/>
