+++ legacy-data-access/src/main/java/gov/ca/cwds/data/legacy/cms/entity/IBaseOutOfHomePlacement.java
@@ -4,7 +4,7 @@ package gov.ca.cwds.data.legacy.cms.entity;
  * @author CWDS CALS API Team
  */
 @SuppressWarnings("squid:S1609") // this interface is not meant to be functional
-public interface IBaseOutOfHomePlacement<T extends BasePlacementHome> {
+public interface IBaseOutOfHomePlacement<T extends PlacementHome> {
 
    T getPlacementHome();
 }
