+++ legacy-data-access/src/test/resources/hibernate.cfg.xml
@@ -62,8 +62,7 @@
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.TribalMembershipVerification" />
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.SpecialProject" />
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.SpecialProjectReferral" />
-   
-
+    <mapping class="gov.ca.cwds.data.legacy.cms.entity.LicensingIssue"/>
 
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ActiveServiceComponentType"/>
     <mapping class="gov.ca.cwds.data.legacy.cms.entity.syscodes.ApprovalStatusType"/>
